Basic Framework
Started a react project in intelliJ with basic npm create template.
Added functionality to call the api for top tracks from a random artist ID in the topArtists.json file and then output the tracks into a simple to table to confirm they were successfully added to the backend.
Also added a new game button that allows for a new artist to be pulled